Move strong types for time to separate files
This can resolve the circular dependencies between `Time.hpp` and
`FramLayout.hpp`.
