+++ Sts1CobcSw/CobcSoftware/EduPowerManagementThread.cpp
@@ -9,6 +9,7 @@
 #include <Sts1CobcSw/Periphery/PersistentState.hpp>
 #include <Sts1CobcSw/Utility/Debug.hpp>
 #include <Sts1CobcSw/Utility/Time.hpp>
+#include <Sts1CobcSw/Utility/TimeTypes.hpp>
 
 #include <strong_type/difference.hpp>
 #include <strong_type/type.hpp>
