+++ Sts1CobcSw/CobcSoftware/TopicsAndSubscribers.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 
-#include <Sts1CobcSw/Utility/Time.hpp>
+#include <Sts1CobcSw/Utility/TimeTypes.hpp>
 
 #include <rodos_no_using_namespace.h>
