+++ Sts1CobcSw/Utility/Time.cpp
@@ -1,5 +1,7 @@
 #include <Sts1CobcSw/Utility/Time.hpp>
 
+#include <strong_type/difference.hpp>
+
 
 namespace sts1cobcsw::internal
 {
