+++ Sts1CobcSw/Utility/Time.ipp
@@ -42,19 +42,4 @@ inline auto SuspendFor(Duration duration) -> void
 {
     RODOS::AT(RODOS::NOW() + value_of(duration));
 }
-
-
-template<std::endian endianness>
-inline auto SerializeTo(void * destination, RealTime const & data) -> void *
-{
-    return SerializeTo<endianness>(destination, value_of(data));
-}
-
-
-template<std::endian endianness>
-inline auto DeserializeFrom(void const * source, RealTime * data) -> void const *
-{
-    source = DeserializeFrom<endianness>(source, &(value_of(*data)));
-    return source;
-}
 }
